+++ knip.config.mjs
@@ -8,6 +8,7 @@ const config = {
 		'node/src/srtpParametersFbsUtils.ts',
 	],
 	ignoreDependencies: ['open-cli', 'supports-color'],
+	ignoreBinaries: ['python3'],
 	typescript: {
 		config: ['tsconfig.json'],
 	},